refactor: deduplicate CollectionData into $lib/types
was defined in 4 places (album, playlist, artist embed routes, and
CollectionEmbed component). now lives in $lib/types.ts once.
